added ship manage actions

+++ src/main/java/com/java/cruisecompany/model/repository/EntityDAO.java
@@ -9,7 +9,7 @@ public interface EntityDAO<T> {
     void create(T entity) throws DAOException;
     void update(T entity) throws DAOException;
     void delete(long id) throws DAOException;
-    Optional <T> findById(int id) throws DAOException;
+    Optional <T> findById(long id) throws DAOException;
     List<T> findAll() throws DAOException;
     List<T> findSorted(String query) throws DAOException;
     long getNumOfRows(String query) throws DAOException;
